Remove the jutsu if it goes off the page
{788: <main.Jutsu object at 0x7fba1ed36d30>, 980: <main.Jutsu object at 0x7fba1ed36ac0>, 176: <main.Jutsu object at 0x7fba1ed36c70>}
print(jutsu)

Allowed for the client to check if the jutsu goes off the screen and send a signal to the server to remove it.
This allows for easier removal of the jutsu by the client in other circumstances as well.

Try except statement likely needed due to data collisions