Refine crimson mist barrier jutsu
As the crimson mist barrier jutsu is so large, it is impractical to simply use the fact that the jutsu has gone outside of the boundaries to remove it as it is likely to still be visible by the player, making it seem like it randomly disappeared. As a result, I decided to use the recently implemented lifetime feature. However, this caused a different error with how the jutsu were removed. What happens when a jutsu has a lifeline and is outside the boundaries? As a result, I had to change the jutsu removal mechanism.